fix(6993): adjust file separations: split utils into dedicated modules, Extract artifacts, bundle-size, performance-benchmarks, and dapp-benchmarks into their own files with colocated tests. Unify all benchmark pipelines and slim index.ts to orchestration.
