+++ development/metamaskbot-build-announce/bundle-size.test.ts
@@ -1,0 +1,45 @@
+import { getHumanReadableSize, getPercentageChange } from './bundle-size';
+
+describe('getHumanReadableSize', () => {
+  it('returns "0 Bytes" for zero', () => {
+    expect(getHumanReadableSize(0)).toBe('0 Bytes');
+  });
+
+  it('formats small values in Bytes', () => {
+    expect(getHumanReadableSize(512)).toBe('512 Bytes');
+  });
+
+  it('formats values over 1024 in KiB', () => {
+    expect(getHumanReadableSize(1536)).toBe('1.5 KiB');
+  });
+
+  it('formats values over 1 MiB in MiB', () => {
+    expect(getHumanReadableSize(2 * 1024 * 1024)).toBe('2 MiB');
+  });
+
+  it('handles negative values', () => {
+    expect(getHumanReadableSize(-2048)).toBe('-2 KiB');
+  });
+});
+
+describe('getPercentageChange', () => {
+  it('returns 0 when both values are zero', () => {
+    expect(getPercentageChange(0, 0)).toBe(0);
+  });
+
+  it('returns 100 when from is zero and to is non-zero', () => {
+    expect(getPercentageChange(0, 500)).toBe(100);
+  });
+
+  it('calculates positive percentage increase', () => {
+    expect(getPercentageChange(100, 150)).toBe(50);
+  });
+
+  it('calculates negative percentage decrease', () => {
+    expect(getPercentageChange(200, 100)).toBe(-50);
+  });
+
+  it('rounds to two decimal places', () => {
+    expect(getPercentageChange(3, 4)).toBe(33.33);
+  });
+});
