+++ shared/constants/benchmarks.ts
@@ -14,3 +14,17 @@ export const BENCHMARK_BUILD_TYPES = {
   BROWSERIFY: 'browserify',
   WEBPACK: 'webpack',
 } as const;
+
+/**
+ * Platform and build-type combinations for which entry benchmarks
+ * (interaction & user journey) are currently collected.
+ * Extend these arrays here to automatically include more combos in
+ * both the PR comment builder and the benchmark gate checks.
+ */
+export const ENTRY_BENCHMARK_PLATFORMS: ReadonlyArray<
+  (typeof BENCHMARK_PLATFORMS)[keyof typeof BENCHMARK_PLATFORMS]
+> = [BENCHMARK_PLATFORMS.CHROME];
+
+export const ENTRY_BENCHMARK_BUILD_TYPES: ReadonlyArray<
+  (typeof BENCHMARK_BUILD_TYPES)[keyof typeof BENCHMARK_BUILD_TYPES]
+> = [BENCHMARK_BUILD_TYPES.BROWSERIFY];
